+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Node: Added ZCOUNT command ([#909](https://github.com/aws/glide-for-redis/pull/909))
 * Python: Added ECHO command ([#953](https://github.com/aws/glide-for-redis/pull/953))
 * Python: Added ZPOPMIN command ([#975](https://github.com/aws/glide-for-redis/pull/975))
+* Node: Added STRLEN command ([#993](https://github.com/aws/glide-for-redis/pull/993))
 
 #### Features
 * Python, Node: Added support in Lua Scripts ([#775](https://github.com/aws/glide-for-redis/pull/775), [#860](https://github.com/aws/glide-for-redis/pull/860))